Renamed DynamicFunction and moved it to hv.util
