+++ holoviews/core/spaces.py
@@ -204,10 +204,10 @@ class HoloMap(UniformNdMapping):
             return self.clone(items, kdims=dimensions, label=self._label, group=self._group)
         elif isinstance(other, self.data_type):
             if isinstance(self, DynamicMap):
-                from .operation import Dynamic
+                from ..util import Dynamic
                 def dynamic_mul(element):
                     return element * other
-                return Dynamic(self, callable=dynamic_mul)
+                return Dynamic(self, operation=dynamic_mul)
             items = [(k, v * other) for (k, v) in self.data.items()]
             return self.clone(items, label=self._label, group=self._group)
         else:
